Tweak minecraft and mouse settings
Turn off particles and head bobbing for performance. I think setting an
FPS cap of 10 could help keep frame times more constant.

I also set the gamemode to creative for safety reasons.

Now for the kicker:
I got annoyed with how often it would pause the same for hitting the
edge of the window.

Set mouse sensitivity to 200% and reduced the mouse movement speed by a
half.

I also made the mouse only move in state==Normal and make it stay in the
center and skip mouse movements. It was having issues with the higher
sensitivity, and I'd rather have it move slower rather than jump quickly
right after the pause.